requirements.txt
# =============================================================================
# Production Dependencies
# =============================================================================

# Web Framework
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0

# Async HTTP client (for forwarding to Zapier)
httpx>=0.27.0,<1.0.0

# Data validation and settings management
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0

# Idempotency cache (TTL-based in-memory deduplication)
cachetools>=5.3.0,<6.0.0

# Structured logging
structlog>=24.1.0,<25.0.0

# .env file loading
python-dotenv>=1.0.0,<2.0.0

# =============================================================================
# Development / Testing Dependencies (install with: pip install -r requirements-dev.txt)
# =============================================================================
# See requirements-dev.txt
